Initial commit for the parsing and handling of EPF persistent damage.
Refactored and edited portions of the EPF_Automation.py file into EPF_resists.py.
